+++ GUI-dev/src/app/@theme/components/tac/tac.component.ts
@@ -1,0 +1,20 @@
+/**
+ * @license
+ * Copyright Akveo. All Rights Reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
+ */
+import { Component } from '@angular/core';
+
+
+
+@Component({
+  selector: 'ngx-eb-tac',
+  templateUrl: 'tac.component.html' ,
+})
+export class EbTacComponent {
+
+  constructor() {
+
+  }
+
+}