+++ GUI-dev/src/app/pages/home/home.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ngx-dashboard',
+  templateUrl: './home.component.html',
+})
+export class HomeComponent {
+}