+++ GUI-dev/src/app/pages/pages-routing.module.ts
@@ -3,6 +3,10 @@ import { NgModule } from '@angular/core';
 
 import { PagesComponent } from './pages.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
+import { HomeComponent } from './home/home.component';
+import {AuthGuard} from '../auth-guard.service';
+import {UserDetailsComponent} from './user-details/user-details.component';
+
 
 const routes: Routes = [{
   path: '',
@@ -11,11 +15,21 @@ const routes: Routes = [{
     {
       path: 'dashboard',
       component: DashboardComponent,
+      canActivate: [AuthGuard],
+    },
+    {
+      path: 'home',
+      component: HomeComponent,
     },
     {
       path: '',
-      redirectTo: 'dashboard',
+      redirectTo: 'home',
       pathMatch: 'full',
+    },
+    {
+      path: 'user-details',
+      component: UserDetailsComponent,
+      canActivate: [AuthGuard],
     },
   ],
 }];