+++ GUI-dev/src/app/pages/pages.module.ts
@@ -1,9 +1,10 @@
 import { NgModule } from '@angular/core';
-
 import { PagesComponent } from './pages.component';
 import { DashboardModule } from './dashboard/dashboard.module';
 import { PagesRoutingModule } from './pages-routing.module';
 import { ThemeModule } from '../@theme/theme.module';
+import {HomeModule} from './home/home.module';
+import {UserDetailsModule} from './user-details/user-details.module';
 
 const PAGES_COMPONENTS = [
   PagesComponent,
@@ -14,6 +15,9 @@ const PAGES_COMPONENTS = [
     PagesRoutingModule,
     ThemeModule,
     DashboardModule,
+    HomeModule,
+    UserDetailsModule,
+
   ],
   declarations: [
     ...PAGES_COMPONENTS,